requirements.txt
flask
flask_cors
tensorflow
numpy
Pillow
gunicorn